feat(intergration/vercel): convert analytics to ts and support in edge
