--- packages/integrations/vercel/analytics.js
@@ -1,63 +1,0 @@
-import { inject } from '@vercel/analytics';
-import { getCLS, getFCP, getFID, getLCP, getTTFB } from 'web-vitals';
-
-const vitalsUrl = 'https://vitals.vercel-analytics.com/v1/vitals';
-
-const getConnectionSpeed = () => {
-  return 'connection' in navigator &&
-    navigator['connection'] &&
-    'effectiveType' in navigator['connection']
-    ? navigator['connection']['effectiveType']
-    : '';
-};
-
-const sendToAnalytics = (metric, options) => {
-  const body = {
-    dsn: options.analyticsId,
-    id: metric.id,
-    page: options.path,
-    href: location.href,
-    event_name: metric.name,
-    value: metric.value.toString(),
-    speed: getConnectionSpeed(),
-  };
-
-  if (options.debug) {
-    console.log('[Analytics]', metric.name, JSON.stringify(body, null, 2));
-  }
-
-  const blob = new Blob([new URLSearchParams(body).toString()], {
-    type: 'application/x-www-form-urlencoded',
-  });
-  if (navigator.sendBeacon) {
-    navigator.sendBeacon(vitalsUrl, blob);
-  } else
-    fetch(vitalsUrl, {
-      body: blob,
-      method: 'POST',
-      credentials: 'omit',
-      keepalive: true,
-    });
-};
-
-function webVitals() {
-  const options = {
-    path: window.location.pathname,
-    analyticsId: import.meta.env.PUBLIC_VERCEL_ANALYTICS_ID,
-  };
-  try {
-    getFID((metric) => sendToAnalytics(metric, options));
-    getTTFB((metric) => sendToAnalytics(metric, options));
-    getLCP((metric) => sendToAnalytics(metric, options));
-    getCLS((metric) => sendToAnalytics(metric, options));
-    getFCP((metric) => sendToAnalytics(metric, options));
-  } catch (err) {
-    console.error('[Analytics]', err);
-  }
-}
-
-const mode = import.meta.env.MODE;
-inject({ mode });
-if (mode === 'production') {
-  webVitals();
-}
